Support newer foreman start syntax
Since 0.35.0, foreman defaults to a concurrency of zero when specifying
concurrency options (https://github.com/ddollar/foreman/pull/132).

Depending on a newer version than this menas that it no longer matters
if entries are added to the Procfile, but not to the Pinfile. This also
allows more of the processes to have coloured output.