Allow excluding a process using the --without option
This adds an OptionParser instance which extracts the -w and --without
option to build a list of processes which should be excluded.

These exclusions are removed from the list of processes to launch after
the dependency tree has been traversed.

There's improved end-to-end testing of the CLI class here too.