Raise error when encountering an unknown process
Currently bowler will not complain when you try to run a process that
isn't defined in the Pinfile. This is sometimes confusing, as other
processes might mask the fact that you misspelled a process name.

For example:

```
$ bundle exec bowl content-api content-store
I, [2015-10-22T22:06:23.944134 #28559]  INFO -- : Starting content-api,
content-store..
22:06:24 content-store.1           | started with pid 28655
22:06:26 content-store.1           | => Booting WEBrick
```

From the output it isn't clear that `content-api` isn't running.

This commit verifies that the process exist. If it doesn't, an error
will be shown:

```
$ bundle exec bowl content-api content-store
E, [2015-10-22T22:10:55.729639 #28920] ERROR -- : Bowler could not load
the Pinfile due to an error: process content-api not found
```