Add --output-only
This flag will make Bowler just output the apps that will be run for a
command.

For example:

```
vagrant@development:~/govuk/development$ bowl collections --output-only
static
content-store
router-api
collections
... (omitted)
```

This allows us to build other scripts that depend on the output. In
particular, we're working on a little script that automatically clones
all dependencies for an application. This will make onboarding new
developers on a large project a bit easier.